feat: codegen translation
