+++ package.json
@@ -7,7 +7,7 @@
     "start": "npm run generateSchema && npm run buildData && npm run translate",
     "generateSchema": "mkdir -p build && quicktype --src types/index.ts --src-lang ts --lang schema --out build/schema-quicktype.json && ts-node scripts/tweak-schema.ts",
     "buildData": "ts-node scripts/build-data.ts",
-    "translate": "ts-node translator/translate.ts",
+    "translate": "ts-node translator/index.ts",
     "tsc": "tsc",
     "build": "npm run generateSchema && npm run buildData && npm run translate && npm run tsc",
     "checkPoi": "ts-node scripts/check-quest.ts",
@@ -43,7 +43,7 @@
     "jest": "^25.2.4",
     "lodash": "^4.17.15",
     "node-fetch": "^2.6.0",
-    "poi-plugin-translator": "^2.3.26",
+    "poi-plugin-translator": "latest",
     "prettier": "^2.0.2",
     "quicktype": "^15.0.220",
     "ts-jest": "^25.3.0",
